+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/score/RootMeanSquareDeviation.java
@@ -4,18 +4,18 @@ package org.rcsb.strucmotif.domain.score;
  * Score a hit by its root-mean-square deviation.
  */
 public class RootMeanSquareDeviation implements Score {
-    private final double score;
+    private final float score;
 
     /**
      * Construct a RMSD instance by wrapping a double value.
      * @param score the computed score
      */
-    public RootMeanSquareDeviation(double score) {
+    public RootMeanSquareDeviation(float score) {
         this.score = score;
     }
 
     @Override
-    public double value() {
+    public float value() {
         return score;
     }
 }
